MOVE: Queues from common to core
