+++ src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/queues/CloudQueue.kt
@@ -13,7 +13,6 @@
 
 package slatekit.core.queues
 
-import slatekit.common.queues.QueueSource
 import slatekit.core.cloud.CloudSupport
 
 /**
