+++ src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/queues/QueueEntryStatus.kt
@@ -1,4 +1,4 @@
-package slatekit.common.queues
+package slatekit.core.queues
 
 import slatekit.common.EnumLike
 import slatekit.common.EnumSupport
