+++ src/lib/kotlin/slatekit-core/src/main/kotlin/slatekit/core/queues/QueueValueConverter.kt
@@ -1,4 +1,4 @@
-package slatekit.common.queues
+package slatekit.core.queues
 
 
 /**
