+++ src/lib/kotlin/slatekit-examples/src/main/kotlin/slatekit/examples/Example_Queue.kt
@@ -20,8 +20,8 @@ package slatekit.examples
 import slatekit.cmds.Command
 import slatekit.results.Try
 import slatekit.results.Success
-import slatekit.common.queues.QueueSourceInMemory
-import slatekit.common.queues.QueueStringConverter
+import slatekit.core.queues.QueueSourceInMemory
+import slatekit.core.queues.QueueStringConverter
 import slatekit.cmds.CommandRequest
 
 //</doc:import_examples>
