+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/QueueSourceTests.kt
@@ -2,7 +2,7 @@ package test.common
 
 import org.junit.Assert
 import org.junit.Test
-import slatekit.common.queues.QueueSourceInMemory
+import slatekit.core.queues.QueueSourceInMemory
 
 class QueueSourceTests {
