+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/setup/WorkerSampleApi.kt
@@ -3,17 +3,16 @@ package test.setup
 import slatekit.apis.*
 import slatekit.apis.Action
 import slatekit.apis.core.Requests
-import slatekit.apis.support.QueueSupport
-import slatekit.common.queues.QueueSource
 import slatekit.common.Source
 import slatekit.common.CommonContext
 import slatekit.common.utils.Random
+import slatekit.core.queues.QueueSource
 import slatekit.results.*
 
 
 @Api(area = "samples", name = "workerqueue", desc = "sample api to integrating workers, queues, apis")
 class WorkerSampleApi(val ctx: CommonContext, val queues:List<QueueSource<String>> = listOf())
-    : QueueSupport, slatekit.apis.Handler {
+    : QueueSupportImpl, slatekit.apis.Handler {
 
     var _lastResult = ""
